+++ Core.lua
@@ -3,7 +3,7 @@ local ADDON_NAME, addon = ...
 _G.ItemFYI = addon
 
 addon.name = ADDON_NAME
-addon.version = "0.2.5"
+addon.version = "0.2.6"
 addon.sessionSkipped = {}
 addon.current = nil
 addon.candidates = {}
@@ -66,6 +66,7 @@ local defaults = {
         curio = true,
         profession = true,
         recipe = true,
+        progress = true,
     },
 }